ZipFileWriter, 傻逼ZipOutputStream的替代者 优化AttrCode效率 Nixim升级了，WIP 删除注解的StringPool序列化方法 InsnNode现有更方便的nodeType来获取类型，而不需要switch code Visitor模式的ASM已可以使用 其他更改   常量对象type  常量池与常量写入器现在共享list   AccessFlag short改char  DummyPrintStream   ConstNamePool经过计算后发现更好的解决方案 (希望时间复杂度挺得住)   ConstantPool只会在写入时intern常量   删除FastSetList   优化AttributeList   优化Mapper写入zip的方式   修复Lexer解析数字的bug   MZF/ZFW支持设置压缩等级   SLR设定CharList初始化大小   Context使用共享byte[]节省内存   Type: char -> byte   字符串加密(1/3)   删除过时的Method构造方法   删除垃圾日志系统   CodeMapper 顺序问题   删除AbstLexer UTF8byte初始化方法   MZF/ZFW支持自定义字符集   AE服务器加长握手超时

+++ roj/reflect/ReflectionUtils.java
@@ -318,7 +318,7 @@ public final class ReflectionUtils {
             case CLASS:
                 return "Object";
             default:
-                StringBuilder s = new StringBuilder(NativeType.toDesc(c));
+                StringBuilder s = new StringBuilder(NativeType.toDesc((byte) c));
                 s.setCharAt(0, Character.toUpperCase(s.charAt(0)));
                 return s.toString();
         }
